CaiFramework/requirements.txt
python-dotenv
openai
rich
mako
wasabi
cai-framework
